Move components to dir
